Change main class name.
