+++ src/main/java/net/comorevi/moneyapi/EventListener.java
@@ -3,22 +3,23 @@ package net.comorevi.moneyapi;
 import cn.nukkit.event.EventHandler;
 import cn.nukkit.event.Listener;
 import cn.nukkit.event.player.PlayerJoinEvent;
+import net.comorevi.moneyapi.util.TextValues;
 
 public class EventListener implements Listener{
 
-    private MoneySAPI plugin;
+    private Main plugin;
 
-    public EventListener(MoneySAPI plugin) {
+    public EventListener(Main plugin) {
         this.plugin = plugin;
     }
 
     @EventHandler
     public void onPlayerJoin(PlayerJoinEvent event){
         if(!plugin.getSQL().existsAccount(event.getPlayer().getName())){
-        	plugin.create(event.getPlayer().getName(), MoneySAPI.defaultmoney);
+        	plugin.create(event.getPlayer().getName(), Main.defaultmoney);
             plugin.getServer().broadcastMessage(TextValues.INFO + plugin.translateString("player-account-add", event.getPlayer().getName()));
         }else{
-            event.getPlayer().sendMessage(TextValues.INFO + plugin.translateString("player-account-load", event.getPlayer().getName(), String.valueOf(plugin.getMoney(event.getPlayer().getName())), MoneySAPI.unit));
+            event.getPlayer().sendMessage(TextValues.INFO + plugin.translateString("player-account-load", event.getPlayer().getName(), String.valueOf(plugin.getMoney(event.getPlayer().getName())), Main.unit));
         }
     }
