+++ src/main/resources/nukkit.yml
@@ -1,6 +1,6 @@
 name: MoneySAPI
-main: net.comorevi.moneyapi.MoneySAPI
-version: 3.1.2
+main: net.comorevi.moneyapi.Main
+version: 4.0.0
 load: POSTWORLD
 author: popkechupki
 api: ["1.0.0"]
